Move Ask-cfpb assets to app
