+++ cfgov/ask_cfpb/jinja2/ask-cfpb/answer-page.html
@@ -1,7 +1,7 @@
 {% extends 'v1/layouts/layout-2-1.html' %}
 {% import 'v1/includes/macros/util/format/datetime.html' as dt %}
 {% import 'v1/includes/templates/streamfield-sidefoot.html' as streamfield_sidefoot with context %}
-{% import 'v1/includes/organisms/ask-search.html' as ask_search with context %}
+{% import './_ask-search.html' as ask_search with context %}
 
 {% block title -%}
     {{- seo_title or page.question or page.title | striptags }} | {{ _('Consumer Financial Protection Bureau') }}
@@ -13,7 +13,7 @@
 
 {% block css -%}
     {{ super() }}
-    <link rel="stylesheet" href="{{ static('css/on-demand/ask.css') }}">
+    <link rel="stylesheet" href="{{ static('apps/ask-cfpb/css/main.css') }}">
 {%- endblock css %}
 
 {% block content_modifiers -%}
@@ -114,3 +114,8 @@
         {% include '_disclaimer.html' %}
     {% endif %}
 {% endblock %}
+
+{% block javascript scoped %}
+    {{ super() }}
+    <script src="{{ static('apps/ask-cfpb/js/main.js') }}"></script>
+{% endblock javascript %}
