+++ cfgov/ask_cfpb/jinja2/ask-cfpb/answer-search-results.html
@@ -1,11 +1,11 @@
 {% import 'v1/includes/molecules/pagination.html' as pagination with context %}
-{% import 'v1/includes/organisms/ask-search.html' as ask_search with context %}
+{% import './_ask-search.html' as ask_search with context %}
 
 {% extends 'v1/layouts/layout-2-1.html' %}
 
 {% block css -%}
     {{ super() }}
-    <link rel="stylesheet" href="{{ static('css/on-demand/ask.css') }}">
+    <link rel="stylesheet" href="{{ static('apps/ask-cfpb/css/main.css') }}">
 {%- endblock css %}
 
 {% block content_modifiers -%}
@@ -101,3 +101,8 @@
         {% include '_disclaimer.html' %}
     {% endif %}
 {%- endblock %}
+
+{% block javascript scoped %}
+    {{ super() }}
+    <script src="{{ static('apps/ask-cfpb/js/main.js') }}"></script>
+{% endblock javascript %}
