+++ cfgov/ask_cfpb/jinja2/ask-cfpb/landing-page.html
@@ -1,12 +1,12 @@
 {% extends 'v1/layouts/layout-2-1.html' %}
 {% import 'v1/includes/templates/render_block.html' as render_block with context %}
-{% import 'v1/includes/organisms/ask-search.html' as ask_search with context %}
+{% import './_ask-search.html' as ask_search with context %}
 
 {% set read_answer_text = _('Read answer') %}
 
 {% block css -%}
     {{ super() }}
-    <link rel="stylesheet" href="{{ static('css/on-demand/ask.css') }}">
+    <link rel="stylesheet" href="{{ static('apps/ask-cfpb/css/main.css') }}">
 {%- endblock css %}
 
 {% block content_modifiers -%}
@@ -45,3 +45,8 @@
         {% include '_disclaimer.html' %}
     {% endif %}
 {%- endblock %}
+
+{% block javascript scoped %}
+    {{ super() }}
+    <script src="{{ static('apps/ask-cfpb/js/main.js') }}"></script>
+{% endblock javascript %}
