+++ cfgov/unprocessed/apps/ask-cfpb/css/main.less
@@ -1,4 +1,5 @@
-@import (reference) '../main.less';
+@import (reference) '../../../css/main.less';
+@import (less) './ask-search.less';
 
 /* ==========================================================================
    consumerfinance.gov
