+++ cfgov/unprocessed/css/main.less
@@ -82,7 +82,6 @@
 
 /* Organism pieces
    ========================================================================== */
-@import (less) 'organisms/ask-search.less';
 @import (less) 'organisms/audio-player.less';
 @import (less) 'organisms/footer.less';
 @import (less) 'organisms/filterable-list-controls.less';
