--- cfgov/unprocessed/js/routes/ask-cfpb/single.js
@@ -1,3 +1,0 @@
-import '../on-demand/ask-autocomplete';
-import '../on-demand/summary';
-import './ask-analytics';
