--- cfgov/unprocessed/js/routes/es/single.js
@@ -1,2 +1,0 @@
-import '../on-demand/ask-autocomplete';
-import '../on-demand/summary';
