+++ cfgov/unprocessed/js/routes/on-demand/summary.js
@@ -1,3 +1,2 @@
 import { Summary } from '@cfpb/cfpb-expandables';
-
 Summary.init();
